+++ CHANGELOG.md
@@ -1,4 +1,2 @@
-- Fix bug where secrets were not attached to functions when using older Functions SDK (#4797).
-- Increase timeout of the Functions Emulator to wait for spawned process to initialize (#4944).
-- Fix bug where Storage Emulator did not convert non-string metadata value as key (#4955).
+- Add v2 Remote Config triggers to deploy (#4937).
 - Allow Firestore emulator UI websocket port to be configurable (#4925).